added addittional commands
