+++ src/temds/cli/preprocess.py
@@ -127,67 +127,5 @@ def cmip6_daily(
         return
     log.info('Preprocess cmip6-daily complete!')
 
-# @app.command()
-# def general_daily(
-#         context: Context,
-#         destination: common.DESTINATION_DIR,
-#         source: common.SOURCE_DIR,
-#         years: Annotated[tuple[int, int], Argument(help="Start and end of years to download data for. Will default to full range of experiment provided")] = None,
-#         source_match: Annotated[str, Option(help=f"")] = '*.nc',
-#         name: Annotated[str, Option(help=f"")] = 'cmpi6',
-
-#         # region_directory: Annotated[Path, Option(help='region folder with manifest.yml, this will supersede the default destination path')]= None,
-
-#         # overwrite: common.OVERWRITE_FLAG = False,
-#         # cleanup: common.CLEANUP_FLAG = False
-#     ):
-#     log = context.obj.log
-#     overwrite = context.obj.overwrite
-#     cleanup = context.obj.cleanup
-#     in_memory = context.obj.in_memory
-#     data = []
-
-#     # files = Path('working/02-arctic/cru-jra-standard/')
-#     full_extent = datasources.timeseries.YearlyTimeSeries(
-#         source, logger=log, in_memory=in_memory
-#     ) 
 
 
-
-
-@app.command()
-def calculate_climate_baseline(
-        context: Context,
-        region_directory: common.DESTINATION_DIR,
-        source: Annotated[str, Argument(help="name of data in region to calculate baseline for")],
-        years: Annotated[tuple[int, int], Argument(help="Start and end of years to download data for. Will default to full range of experiment provided")] = None,
-        name: Annotated[str, Argument(help=f"name to save baseline data in region to; When not provided -baseline is appended to source")] = None,
-
-    ):
-    """Preprocesses downloaded ERA5 daily data. Preprocessed data will be
-    formatted to be read as a YearlyDataset.
-    """
-    log = context.obj.log
-    overwrite = context.obj.overwrite
-    cleanup = context.obj.cleanup
-
-    if context.obj.region: 
-        log.info('Using region from context')
-        area = context.obj.region
-        region_directory = context.obj.region_directory
-    else:
-        log.info('Using region from argument')
-        area = Region.from_directory(region_directory, logger=log)
-
-    if name is None:
-        name = source + '-baseline'
-
-    area.calculate_climate_baseline(years[0], years[1], name, source)
-
-    if context.obj.save_enabled:
-        try:
-            area.export_to_directory(region_directory, items=[name], update_manifest=True, overwrite=overwrite)
-        except FileExistsError:
-            log.error('Output files exist. Cannot save unless --overwrite is passed.')
-            sys.exit(0)
-    return area
